+++ apps/components/ReactifiedComponentsWrapper.js
@@ -21,6 +21,17 @@ function clickerDiClick() {
 	console.log( "You are an exceptional clicker!" );
 }
 
+const buttonRef = React.createRef();
+const buttonStyledLinkRef = React.createRef();
+
+const focusButtonRef = () => {
+	buttonRef.current.focus();
+};
+
+const focusLinkRef = () => {
+	buttonStyledLinkRef.current.focus();
+};
+
 const buttonGrouping = <Fragment>
 	<h3>"primary" variant (default)</h3>
 	<Button onClick={ clickerDiClick } title="Testing whether other props are also passed, like this tooltip">Default button</Button>
@@ -39,6 +50,12 @@ const buttonGrouping = <Fragment>
 	<ButtonStyledLink variant="secondary" href={ "#" }>Secondary link</ButtonStyledLink>
 	<ButtonStyledLink variant="secondary" small={ true } href={ "#" }>Secondary small link</ButtonStyledLink>
 
+	<h3>With Ref!</h3>
+	<Button variant="secondary" buttonRef={ buttonRef } small={ true } onClick={ clickerDiClick }>This Button has a Ref!</Button>
+	<button onClick={ focusButtonRef }>focus test</button>
+	<ButtonStyledLink variant="secondary" buttonRef={ buttonStyledLinkRef } small={ true } href={ "#" }>This ButtonStyledLink has a Ref!</ButtonStyledLink>
+	<button onClick={ focusLinkRef }>focus test</button>
+
 	<h3>"buy" variant (or "upsell")</h3>
 	<Button variant="upsell" onClick={ clickerDiClick }>Buy button</Button>
 	<Button variant="upsell" disabled={ true } onClick={ clickerDiClick }>Buy disabled button</Button>
@@ -52,6 +69,10 @@ const buttonGrouping = <Fragment>
 	<Button variant="remove" onClick={ clickerDiClick }>Remove button</Button>
 	<ButtonStyledLink variant="hide" href="#">Hide Link</ButtonStyledLink>
 	<ButtonStyledLink variant="remove" href="#">Remove Link</ButtonStyledLink>
+
+	<h3>"edit" variant</h3>
+	<Button variant="edit" onClick={ clickerDiClick }>Edit button</Button>
+	<ButtonStyledLink variant="edit" href="#">Edit Link</ButtonStyledLink>
 
 	<h3>CloseButton (not a variant due its different characteristics)</h3>
 	<CloseButton onClick={ clickerDiClick } />
@@ -182,10 +203,10 @@ const ReactifiedComponentsWrapper = () => {
 				name="my-selection"
 				options={
 					[
-						{ name: "option 1", value: "opt1" },
-						{ name: "option 2", value: "opt2" },
-						{ name: "option 3", value: "opt3" },
-						{ name: "option 4", value: "opt4" },
+						{ name: "Option 1", value: "opt1" },
+						{ name: "Option 2", value: "opt2" },
+						{ name: "Option 3", value: "opt3" },
+						{ name: "Option 4", value: "opt4" },
 					]
 				}
 				selected={ [ "opt1", "opt3" ] }