+++ packages/yoastseo/spec/assessments/sentenceLengthInTextSpec.js
@@ -387,6 +387,37 @@ describe( "An assessment for sentence length", function() {
 		expect( assessment.hasMarks() ).toBe( false );
 	} );
 
+	it( "returns the score for 100% long sentences in Arabic", function() {
+		const mockPaper = new Paper( "text", { locale: "ar_AR" } );
+		const sentenceLengthInTextAssessmentArabic = new SentenceLengthInTextAssessment( contentConfiguration( mockPaper.getLocale() ).sentenceLength );
+
+		const assessment = sentenceLengthInTextAssessmentArabic.getResult( mockPaper, Factory.buildMockResearcher( [
+			{ sentence: "", sentenceLength: 25 },
+		] ), i18n );
+
+		expect( assessment.hasScore() ).toBe( true );
+		expect( assessment.getScore() ).toEqual( 3 );
+		expect( assessment.getText() ).toEqual( "<a href='https://yoa.st/34v' target='_blank'>Sentence length</a>: " +
+			"100% of the sentences contain more than 20 words, which is more than the recommended maximum of 25%." +
+			" <a href='https://yoa.st/34w' target='_blank'>Try to shorten the sentences</a>." );
+		expect( assessment.hasMarks() ).toBe( true );
+	} );
+
+	it( "returns the score for all short sentences in Arabic", function() {
+		const mockPaper = new Paper( "text", { locale: "ar_AR" } );
+		const sentenceLengthInTextAssessmentArabic = new SentenceLengthInTextAssessment( contentConfiguration( mockPaper.getLocale() ).sentenceLength );
+
+		const assessment = sentenceLengthInTextAssessmentArabic.getResult( mockPaper, Factory.buildMockResearcher( [
+			{ sentence: "", sentenceLength: 16 },
+
+		] ), i18n );
+
+		expect( assessment.hasScore() ).toBe( true );
+		expect( assessment.getScore() ).toEqual( 9 );
+		expect( assessment.getText() ).toEqual( "<a href='https://yoa.st/34v' target='_blank'>Sentence length</a>: Great!" );
+		expect( assessment.hasMarks() ).toBe( false );
+	} );
+
 	it( "returns the score for 100% long sentences in Hebrew", function() {
 		const mockPaper = new Paper( "text", { locale: "he_IL" } );
 		const sentenceLengthInTextAssessmentHebrew = new SentenceLengthInTextAssessment( contentConfiguration( mockPaper.getLocale() ).sentenceLength );