+++ packages/yoastseo/spec/researches/buildTopicStemsSpec.js
@@ -365,6 +365,26 @@ describe( "A test for filtering function words in supported languages", function
 	);
 } );
 
+describe( "A test for filtering function words in supported languages", function() {
+	// Function word: دو
+	const forms = buildStems( "دو سیب", "fa", false );
+	expect( forms ).toEqual(
+		new TopicPhrase(
+			[ new StemOriginalPair( "سیب", "سیب" ) ],
+			false )
+	);
+} );
+
+describe( "A test for filtering function words in supported languages", function() {
+	// Function word: هذه
+	const forms = buildStems( "هذه المعلومات", "ar", false );
+	expect( forms ).toEqual(
+		new TopicPhrase(
+			[ new StemOriginalPair( "المعلومات", "المعلومات" ) ],
+			false )
+	);
+} );
+
 describe( "A test for topic phrase objects", function() {
 	const testTopicPhrase = new TopicPhrase(
 		[