+++ packages/yoastseo/spec/specHelpers/getMorphologyData.js
@@ -7,6 +7,8 @@ import ru from "../../premium-configuration/data/morphologyData-ru-v9.json";
 import it from "../../premium-configuration/data/morphologyData-it-v9.json";
 import pt from "../../premium-configuration/data/morphologyData-pt-v9.json";
 import id from "../../premium-configuration/data/morphologyData-id-v9.json";
+import pl from "../../premium-configuration/data/morphologyData-pl-v9.json";
+import ar from "../../premium-configuration/data/morphologyData-ar-v9.json";
 
 
 const morphologyData = {
@@ -19,6 +21,8 @@ const morphologyData = {
 	it,
 	pt,
 	id,
+	pl,
+	ar,
 };
 
 /**