+++ packages/yoastseo/spec/stringProcessing/matchWordInSentenceSpec.js
@@ -41,3 +41,20 @@ describe( "returns whether a word is in the sentence", function() {
 		expect( isWordInSentence( word, sentence ) ).toEqual( expected );
 	} );
 } );
+
+describe( "returns whether a word is in the sentence in RTL languages with special punctuation", function() {
+	it( "returns true if the word is followed by Urdu-specific punctuation mark", function() {
+		const word = "گئے";
+		const sentence = "اس دوران میں وہ حملے کرتے رہے اور آخرکار شکست کھا گئے۔";
+		const expected = true;
+		expect( isWordInSentence( word, sentence ) ).toEqual( expected );
+	} );
+
+	it( "returns true if the word is followed by an Arabic-specific punctuation mark", function() {
+		const word = "الجيدة";
+		const sentence = "ما هي المقالات الجيدة؟";
+		const expected = true;
+		expect( isWordInSentence( word, sentence ) ).toEqual( expected );
+	} );
+} );
+