+++ packages/yoastseo/src/helpers/getFunctionWords.js
@@ -28,6 +28,10 @@ import indonesianFunctionWordsFactory from "../researches/indonesian/functionWor
 const indonesianFunctionWords = indonesianFunctionWordsFactory();
 import hebrewFunctionWordsFactory from "../researches/hebrew/functionWords.js";
 const hebrewFunctionWords = hebrewFunctionWordsFactory();
+import arabicFunctionWordsFactory from "../researches/arabic/functionWords.js";
+const arabicFunctionWords = arabicFunctionWordsFactory();
+import farsiFunctionWordsFactory from "../researches/farsi/functionWords.js";
+const farsiFunctionWords = farsiFunctionWordsFactory();
 
 /**
  * Returns the function words for all languages.
@@ -48,5 +52,7 @@ export default function() {
 		sv: swedishFunctionWords,
 		id: indonesianFunctionWords,
 		he: hebrewFunctionWords,
+		ar: arabicFunctionWords,
+		fa: farsiFunctionWords,
 	};
 }