+++ packages/yoastseo/src/helpers/getStemForLanguage.js
@@ -7,6 +7,8 @@ import russianDetermineStem from "../morphology/russian/stem";
 import italianDetermineStem from "../morphology/italian/stem";
 import portugueseDetermineStem from "../morphology/portuguese/stem";
 import indonesianDetermineStem from "../morphology/indonesian/stem";
+import polishDetermineStem from "../morphology/polish/stem";
+import arabicDetermineStem from "../morphology/arabic/stem";
 
 /**
  * Collects all functions for determining a stem per language and returns this collection to a Researcher
@@ -24,5 +26,7 @@ export default function() {
 		it: italianDetermineStem,
 		pt: portugueseDetermineStem,
 		id: indonesianDetermineStem,
+		pl: polishDetermineStem,
+		ar: arabicDetermineStem,
 	};
 }