+++ packages/yoastseo/src/helpers/getTransitionWords.js
@@ -55,6 +55,9 @@ import twoPartTransitionWordsHebrew from "../researches/hebrew/twoPartTransition
 
 import getLanguage from "./getLanguage.js";
 
+import transitionWordsArabicFactory from "../researches/arabic/transitionWords.js";
+const transitionWordsArabic = transitionWordsArabicFactory().allWords;
+import twoPartTransitionWordsArabic from "../researches/arabic/twoPartTransitionWords.js";
 
 /**
  * Returns transition words for a specific locale.
@@ -129,6 +132,11 @@ export default function( locale ) {
 			return {
 				transitionWords: transitionWordsHebrew,
 				twoPartTransitionWords: twoPartTransitionWordsHebrew,
+			};
+		case "ar":
+			return {
+				transitionWords: transitionWordsArabic,
+				twoPartTransitionWords: twoPartTransitionWordsArabic,
 			};
 		default:
 		case "en":