+++ packages/yoastseo/src/researches/passiveVoice/morphological/determinePassiveSentence.js
@@ -12,6 +12,9 @@ const passivePrefixIndonesian = "di";
 import nonPassivesIndonesianFactory from "../../indonesian/passiveVoice/nonPassiveVerbsStartingDi";
 const nonPassivesIndonesian = nonPassivesIndonesianFactory();
 
+import getPassiveVerbsArabicFactory from "../../arabic/passiveVoice/passiveVerbsWithLongVowel";
+const getPassiveVerbsArabic = getPassiveVerbsArabicFactory();
+
 /**
  * Matches the sentence against passive verbs.
  *
@@ -82,6 +85,35 @@ const determineSentenceIsPassiveIndonesian = function( sentence ) {
 };
 
 /**
+ * Checks the passed sentence to see if it contains Arabic passive verb-forms.
+ *
+ * @param {string} sentence The sentence to match against.
+ * @returns {Boolean} Whether the sentence contains Arabic passive voice.
+ */
+const determineSentenceIsPassiveArabic = function( sentence ) {
+	const arabicPrepositionalPrefix =  "و";
+	const words = getWords( sentence );
+	const passiveVerbs = [];
+
+	for ( let word of words ) {
+		// Check if the word starts with prefix و
+		if ( word.startsWith( arabicPrepositionalPrefix ) ) {
+			word = word.slice( 1 );
+		}
+		let wordWithDamma = -1;
+		// Check if the first character has a damma or if the word is in the list of Arabic passive verbs
+		if ( word.length >= 2 ) {
+			wordWithDamma = word[ 1 ].search( "\u064F" );
+		}
+		if ( wordWithDamma !== -1 || getPassiveVerbsArabic.includes( word ) ) {
+			passiveVerbs.push( word );
+		}
+	}
+
+	return passiveVerbs.length !== 0;
+};
+
+/**
  * Determines whether a sentence is passive.
  *
  * @param {string} sentenceText The sentence to determine voice for.
@@ -96,5 +128,9 @@ export default function( sentenceText, language ) {
 
 	if ( language === "id" ) {
 		return determineSentenceIsPassiveIndonesian( sentenceText, language );
+	}
+
+	if ( language === "ar" ) {
+		return determineSentenceIsPassiveArabic( sentenceText, language );
 	}
 }